Storyboard Templates
These Templates can help with get a general design look and feel for the
webapp, they include the common browsers.

The idea is to just get the basic design down, don't get stuck on
colours, font sizes, or content, just a layout to see how it feels in
the browser without having to stuff around with code.

Sometimes they can be printed out and scanned back in or worked
digitally.